fix: correct data types in wishlist entity and migration for FK compatibility
- Changed userId and productId columns from BIGINT to INT to match users and products table ID types
- Updated Wishlist entity to use 'int' type decorator instead of 'bigint'
- Fixed foreign key constraint compatibility issue
- Created wishlists table successfully in the database

+++ nestapi/src/entities/wishlist.entity.ts
@@ -15,10 +15,10 @@ export class Wishlist {
   @PrimaryGeneratedColumn()
   id: number
 
-  @Column({ type: 'bigint' })
+  @Column({ type: 'int' })
   userId: number
 
-  @Column({ type: 'bigint' })
+  @Column({ type: 'int' })
   productId: number
 
   @ManyToOne(() => User, { onDelete: 'CASCADE' })
